Implemented a published model converter
The purpose of the converter is that the value type of properties using this converter will be a published model (when using ModelsBuilder) rather than the normal "IPublishedContent". If a common type can not be calculated, "IPublishedContent" will still be used as fall back for the value type.

Part of the implementation is borrowed from @callumbwhyte's Super Value Converters package - thanks for originally writing this 👍