Misc changes to how the item converter selection is saved
The Umbraco 10 version currently saves a string, which technically should be fine with what we're doing now, but going forward, I may introduce additonal options to the item converter, in which case it makes sense to save an object instead of just a string value. String values - e.g. from an upgrade - are still supported and won't result in a breaking change.